requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coverage==7.2.5
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2.1
django-debug-toolbar==4.1.0
django-filter==23.2
djangorestframework==3.14.0
idna==3.4
Markdown==3.4.3
oauthlib==3.2.2
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.7.0
python3-openid==3.2.0
pytz==2023.3
requests==2.30.0
requests-oauthlib==1.3.1
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
urllib3==2.0.2
